Factor out more X11-independent code
Factor out daemon-specific code to daemon.c
and poll interface to poller.c.

+++ window-x11.h
@@ -131,7 +131,7 @@ struct window {
 
     struct term *term;
     struct render_cell_state rcstate;
-    size_t poll_index;
+    int poll_index;
 
     struct title_stack_item *title_stack;
